Applying patch from Dennis van Dok

+++ src/services/a-rex/lrms/submit_common.sh
@@ -513,6 +513,7 @@ move_files_to_node () {
       [ "$f" = "$RUNTIME_JOB_DIR/.." ] && continue
       [ "$f" = "$RUNTIME_JOB_DIR/.diag" ] && continue
       [ "$f" = "$RUNTIME_JOB_DIR/.comment" ] && continue
+      [ -f "$f" ] || continue
       if ! $RUNTIME_LOCAL_SCRATCH_MOVE_TOOL "$f" "$RUNTIME_NODE_JOB_DIR"; then
         echo "Failed to '$RUNTIME_LOCAL_SCRATCH_MOVE_TOOL' '$f' to '$RUNTIME_NODE_JOB_DIR'" 1>&2
         exit 1
@@ -643,6 +644,7 @@ move_files_to_frontend () {
         [ "$f" = "$RUNTIME_NODE_JOB_DIR/.." ] && continue
         [ "$f" = "$RUNTIME_NODE_JOB_DIR/.diag" ] && continue
         [ "$f" = "$RUNTIME_NODE_JOB_DIR/.comment" ] && continue
+        [ -f "$f" ] || continue
         if ! mv "$f" "$RUNTIME_FRONTEND_JOB_DIR"; then
           echo "Failed to move '$f' to '$RUNTIME_FRONTEND_JOB_DIR'" 1>&2
           RESULT=1
